- abstracts the snippet model for reusability
